@@ -51,6 +51,7 @@
   },
   "dependencies": {
     "@wecom/jssdk": "^1.3.1",
+    "axios": "^1.4.0",
     "electron-log": "^4.4.8",
     "electron-packager": "^17.1.1",
     "electron-reload": "^2.0.0-alpha.1",
@@ -58,9 +59,12 @@
     "element-plus": "^2.3.4",
     "fs-extra": "^11.1.1",
     "lodash.isequal": "^4.5.0",
+    "mqtt": "^4.3.7",
     "neroom-web-sdk": "^1.14.1",
-    "nertc-electron-sdk": "4.6.503-rc-984",
+    "nertc-web-sdk": "^4.6.50",
+    "node-machine-id": "^1.1.12",
     "pinia": "^2.0.35",
+    "sha1": "^1.1.1",
     "update-electron-app": "^2.0.1",
     "vue": "^3.2.47",
     "vue-router": "^4.1.6"
@@ -76,7 +80,7 @@
     "@vue/test-utils": "^2.3.2",
     "@vue/tsconfig": "^0.3.2",
     "cypress": "^12.11.0",
-    "electron": "^16.2.8",
+    "electron": "^24.3.0",
     "electron-builder": "^23.6.0",
     "eslint": "^8.39.0",
     "eslint-plugin-cypress": "^2.13.3",
@@ -91,6 +95,7 @@
     "vite": "^4.3.3",
     "vite-plugin-resolve": "^2.3.0",
     "vitest": "^0.30.1",
-    "vue-tsc": "^1.6.2"
+    "vue-tsc": "^1.6.2",
+    "vue-wxlogin": "^1.0.4"
   }
 }
