@@ -1,5 +1,5 @@
-import { createApp } from "vue";
-import { createPinia } from "pinia";
+import {createApp} from "vue";
+import {createPinia} from "pinia";
 import "element-plus/theme-chalk/index.css";
 import App from "./App.vue";
 import router from "./router";
