@@ -4,11 +4,13 @@
             <div class="right-side">
                 <div class="doc">
                     <div class="title alt">您可以点击的按钮测试功能</div>
+                    <el-button type="primary" round @click="getSystemInfo">获取系统信息</el-button>
                     <el-button type="primary" round @click="CheckUpdate">检查更新, 不可用于开发环境</el-button>
                     <!-- <el-progress :percent="this.percent" v-show="show">更新进度</el-progress> -->
                     <el-button type="primary" round @click="downloadUpdate">手动下载更新文件</el-button>
                     <el-progress :text-inside="true" :stroke-width="24" :percentage="percent"></el-progress>
-
+                    <el-button type="primary" round @click="changeTheme('light')">亮色</el-button>
+                    <el-button type="primary" round @click="changeTheme('dark')">暗色</el-button>
 
                 </div>
             </div>
@@ -29,7 +31,13 @@ console.log(route.query);
 console.log(window.location.href);
 const percent = ref(0);
 const show = ref(false);
+const getSystemInfo = () => {
+    ipcRenderer.send('getSystemInfo');
+}
 // 主进程返回检查状态
+ipcRenderer.on('getSystemInfo', (e, data) => {
+    console.log(data);
+})
 ipcRenderer.on("message", (e, data) => {
     console.log("status--->", data.status);
     switch (data.status) {
@@ -139,28 +147,31 @@ ipcRenderer.on('downloadProgress', (event, progressObj) => {
 const downloadUpdate = () => {
     console.log('2222');
     ElMessageBox.confirm("下载完成，是否立即更新", "提示", {
-            closeOnClickModal: false, // 禁止点击遮罩关闭弹框
-            closeOnPressEscape: false, // 禁止按 ECS 键关闭弹框
-            confirmButtonText: '确定',
-            cancelButtonText: '取消',
-            type: 'warning',
-        }).then(() => {
-            ipcRenderer.send('handleUpdateNow');
-            ElMessage({
-                type: 'success',
-                message: '更新成功'
-            })
-        }).catch(() => {
-            ElMessage({
-                type: 'info',
-                message: '取消更新'
-            })
+        closeOnClickModal: false, // 禁止点击遮罩关闭弹框
+        closeOnPressEscape: false, // 禁止按 ECS 键关闭弹框
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning',
+    }).then(() => {
+        ipcRenderer.send('handleUpdateNow');
+        ElMessage({
+            type: 'success',
+            message: '更新成功'
         })
+    }).catch(() => {
+        ElMessage({
+            type: 'info',
+            message: '取消更新'
+        })
+    })
     // ipcRenderer.send("downloadUpdate")
 }
 const CheckUpdate = () => {
     console.log('1111');
     ipcRenderer.send("checkForUpdate")
+}
+const changeTheme = (color: string) => {
+    ipcRenderer.send('changeTheme', color)
 }
 </script>
 
